Fix RangeError on too many players
Problem: upon clicking the "add player" button, we correctly don't
add more than the configured maximum permitted number of players,
but then upon clicking on a player to switch the pressed state, we
get a RangeError bug displayed.

Solution: apparently we increase the player counter and only then
check the constraint with an assert statement. Moved the assert to
a correct position.